move RunMetrics code into own component; style tweaks
