+++ src/components/experiment-tracking/accordion/accordion.scss
@@ -77,7 +77,7 @@ $toggle-width: 1.15em;
   display: none;
 }
 
-// Layout with left toggle
+// Layout with left-side toggle
 
 .accordion--left .accordion__heading {
   justify-content: flex-start;
