+++ src/components/experiment-tracking/run-metadata/run-metadata.scss
@@ -2,6 +2,7 @@
   background-color: var(--color-bg-1);
   display: flex;
   flex-shrink: 0;
+  min-width: 100%;
   padding: 5em 0 0 5em;
   width: max-content;
