+++ src/components/experiment-tracking/run-metrics/run-metrics.scss
@@ -1,0 +1,66 @@
+.details-stats {
+  display: flex;
+  flex-shrink: 0;
+  padding: 5em 0 0 5em;
+
+  &--single {
+    padding: 5em 0 0 9em;
+  }
+}
+
+.run {
+  flex-grow: 1;
+  flex-shrink: 0;
+  margin-right: 13em;
+}
+
+.run-details__accordion {
+  margin-bottom: 36px;
+  max-width: 75em;
+  width: 100%;
+}
+
+.run-details__accordion-header {
+  margin-bottom: 16px;
+}
+
+.metric-data-wrapper {
+  display: flex;
+  flex-direction: column;
+  font-size: 1.4em;
+  vertical-align: top;
+}
+
+.row-labels,
+.row-values {
+  display: flex;
+}
+
+.metric-label,
+.metric-value {
+  margin-bottom: 8px;
+}
+
+.metric-label {
+  display: block;
+  width: 230px;
+
+  &--single {
+    width: 50%;
+  }
+}
+
+.metric-value {
+  display: block;
+  width: 310px;
+
+  &--single {
+    width: 50%;
+  }
+}
+
+.metric-label--small,
+.metric-value--small {
+  color: var(--color-text-faded);
+  font-size: 13px;
+}
